refactor: Move sniff under connectivity
The sniff package is used only by the connectivity package. Move it from
utils/sniff to connectivity/sniff to simplify the package structure.
