+++ connectivity/sniff/filters.go
