+++ connectivity/tests/encryption.go
@@ -12,8 +12,8 @@ import (
 	"github.com/cilium/cilium/pkg/versioncheck"
 
 	"github.com/cilium/cilium-cli/connectivity/check"
+	"github.com/cilium/cilium-cli/connectivity/sniff"
 	"github.com/cilium/cilium-cli/utils/features"
-	"github.com/cilium/cilium-cli/utils/sniff"
 )
 
 type requestType int
